#1 Initial typescript state
* Transferring project from js to typescript
- lesson 1-30 -
