--- .idea/dictionaries/samar.xml
@@ -1,7 +1,0 @@
-<component name="ProjectDictionaryState">
-  <dictionary name="samar">
-    <words>
-      <w>navbar</w>
-    </words>
-  </dictionary>
-</component>
