--- src/App.module.css
@@ -1,16 +1,0 @@
-.appWrapper {
-    display: grid;
-    width: 1200px;
-    margin: 0 auto;
-
-    grid-template-areas: "header header" "nav content";
-
-    grid-template-rows: 60px 1fr;
-    grid-template-columns: 2fr 10fr;
-    grid-gap: 10px;
-}
-
-.appWrapperContent {
-    grid-area: content;
-    background-color: darksalmon;
-}
