--- src/components/dialogs/Dialogs.module.css
@@ -1,12 +1,0 @@
-.dialogs {
-    display: grid;
-    grid-template-columns: 2fr 10fr;
-}
-
-.dialogItems {
-    padding: 10px;
-}
-
-.messageItems {
-    padding: 10px;
-}
