--- src/components/header/Header.module.css
@@ -1,8 +1,0 @@
-.header {
-    grid-area: header;
-    background-color: royalblue;
-}
-
-.header img {
-    width: 60px;
-}
