--- src/components/music/Music.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import style from './Music.module.css';
-
-class Music extends React.Component {
-    render = () => {
-        return (
-            <div>
-                Your music should be here
-            </div>
-        );
-    }
-}
-
-export default Music;
