--- src/components/navbar/Navbar.module.css
@@ -1,14 +1,0 @@
-.nav {
-    grid-area: nav;
-    background-color: royalblue;
-    padding: 15px;
-}
-
-.item a {
-    color: white;
-    text-decoration: none;
-}
-
-.item a.activeLink {
-    color: gold;
-}
