--- src/components/news/News.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import style from './News.module.css';
-
-class News extends React.Component {
-    render = () => {
-        return (
-            <div>
-                Your news should be here
-            </div>
-        );
-    }
-}
-
-export default News;
