--- src/components/profile/Profile.module.css
@@ -1,7 +1,0 @@
-.backgroundPhoto {
-}
-
-.backgroundPhoto img {
-    position: relative;
-    width: 100%;
-}
