--- src/components/profile/posts/Posts.module.css
@@ -1,7 +1,0 @@
-.posts {
-    padding: 10px;
-}
-
-.postItems {
-    margin-top: 10px;
-}
