--- src/components/profile/posts/postItem/PostItem.module.css
@@ -1,9 +1,0 @@
-.item {
-    color: gold;
-}
-
-.item img {
-    height: 30px;
-    width: 30px;
-    border-radius: 30px;
-}
