--- src/components/profile/profileInfo/ProfileInfo.module.css
@@ -1,8 +1,0 @@
-.avatar {
-    height: 200px;
-    width: 200px;
-}
-
-.profileDescription {
-    padding: 10px;
-}
