--- src/components/settings/Settings.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import style from './Settings.module.css';
-
-class Settings extends React.Component {
-    render = () => {
-        return (
-            <div>
-                Your settings should be here
-            </div>
-        );
-    }
-}
-
-export default Settings;
